+++ Backend/src/routes.ts
@@ -6,6 +6,7 @@ import emailsRoutes from '@modules/emails/email.routes.js';
 import foldersRoutes from '@modules/folders/folder.routes.js';
 import usersRoutes from '@modules/user/user.routes.js';
 import utilsRoutes from '@modules/utils/index.js';
+import attachmentRoutes from '@modules/attachments/attachment.routes.js';
 
 const router = Router();
 
@@ -22,6 +23,8 @@ router.use('/accounts', accountsRoutes);
 router.use('/emails', emailsRoutes);
 
 router.use('/folders', foldersRoutes);
+
+router.use('/attachments', attachmentRoutes);
 
 router.use('/utils', utilsRoutes);